+++ integration-tests/src/test/java/com/ebsolutions/papertrail/financialdataproviderservice/user/UserSteps.java
@@ -1,9 +1,9 @@
 package com.ebsolutions.papertrail.financialdataproviderservice.user;
 
 import com.ebsolutions.papertrail.financialdataproviderservice.BaseStep;
+import com.ebsolutions.papertrail.financialdataproviderservice.config.TestConstants;
 import com.ebsolutions.papertrail.financialdataproviderservice.model.User;
-import com.ebsolutions.papertrail.financialdataproviderservice.tooling.TestConstants;
-import com.ebsolutions.papertrail.financialdataproviderservice.tooling.UserTestData;
+import com.ebsolutions.papertrail.financialdataproviderservice.testdata.UserTestData;
 import com.ebsolutions.papertrail.financialdataproviderservice.util.ApiCallTestUtil;
 import com.ebsolutions.papertrail.financialdataproviderservice.util.UserTestUtil;
 import com.fasterxml.jackson.core.JsonProcessingException;
