+++ integration-tests/src/test/java/com/ebsolutions/papertrail/financialdataproviderservice/util/CommonTestUtil.java
@@ -1,6 +1,6 @@
 package com.ebsolutions.papertrail.financialdataproviderservice.util;
 
-import com.ebsolutions.papertrail.financialdataproviderservice.tooling.TestConstants;
+import com.ebsolutions.papertrail.financialdataproviderservice.config.TestConstants;
 
 public class CommonTestUtil {
   public static String isEmptyString(String value) {
